requirements.txt
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
click>=8.0.0
plotly>=5.0.0
numpy>=1.24.0
google-genai>=0.3.0
anthropic>=0.34.0
firebase-admin>=6.5.0
Pillow>=10.0.0
python-multipart>=0.0.6